Fix: File sorting function
Function for file sorting moved out of Dashboard.tsx and into
a util file. Sorting now accounts for file names that begin with any
number of digits then any number of or no other characters following.
Function also properly accounts for the number of bytes being represented
in the sortBy property as a number.