Migrated project to use access with Drive add/show
This commit adds the base for the access features to be added. Previously,
it was assumed that the only drives you should see are those that hold your
userId on the record in Drives. Creating a Drive now also creates a bucketsOnUsers
record for which the user is considered the creator of the drive. Drives shown on the
homepage are currently the same as before the commit (those created by the user), but
a separate section will be made for those shared with the user in later commits.