Added: Altering the Drive UI based on User Role
The UI components which will result in writing actions to the drive
currently open will now react to the value of provided RoleContext passed
from a resulting server query. Indicators for the role of the user in the open
drive and tooltips for 'disabled' buttons/components will also be added in
later commits.