Addition of drive type with discriminator for keys
Added a StorageDrive type which will be passed to the client instead of
passing a Model Drive object which would contain all access and secret keys.
The type will allow for both the drive.type and drive.permissions to be valid
discriminators for use within the useBucket hooks and providing different functionality
given the drive is either owned or shared with the user who has opened it.